Organizes tests into package-specific files
Moves existing test cases from a monolithic main_test.go into their
respective pkg subdirectories.

Each pkg now contains its dedicated `_test.go` file for tests and a
`suite_test.go` file for test suite setup. This improves test
organization and maintainability.
